Flat admin cracked

+++ modules/flatadmintheme/core/Apiinit.php
@@ -2,9 +2,8 @@
 
 namespace modules\flatadmintheme\core;
 
-require_once __DIR__ . '/../third_party/node.php';
-require_once __DIR__ . '/../vendor/autoload.php';
-
+require_once __DIR__.'/../third_party/node.php';
+require_once __DIR__.'/../vendor/autoload.php';
 use Firebase\JWT\JWT as Flatadmintheme_JWT;
 use Firebase\JWT\Key as Flatadmintheme_Key;
 use WpOrg\Requests\Requests as Flatadmintheme_Requests;
@@ -14,16 +13,16 @@ class Apiinit
     public static function the_da_vinci_code($module_name)
     {
         $module = get_instance()->app_modules->get($module_name);
-        $verification_id = !empty(get_option($module_name . '_verification_id')) ? base64_decode(get_option($module_name . '_verification_id')) : '';
-        $token = get_option($module_name . '_product_token');
+        $verification_id =  !empty(get_option($module_name.'_verification_id')) ? base64_decode(get_option($module_name.'_verification_id')) : '';
+        $token = get_option($module_name.'_product_token');
 
-        $id_data = explode('|', $verification_id);
-        $verified = !((empty($verification_id)) || (4 != \count($id_data)));
+        $id_data         = explode('|', $verification_id);
+        $verified        = !((empty($verification_id)) || (4 != \count($id_data)));
 
         if (4 === \count($id_data)) {
             $verified = !empty($token);
             try {
-                $data = Flatadmintheme_JWT::decode($token, new Flatadmintheme_Key($id_data[3], 'HS512'));
+                $data =Flatadmintheme_JWT::decode($token, new Flatadmintheme_Key($id_data[3], 'HS512'));
                 if (!empty($data)) {
                     $verified = basename($module['headers']['uri']) == $data->item_id && $data->item_id == $id_data[0] && $data->buyer == $id_data[2] && $data->purchase_code == $id_data[3];
                 }
@@ -31,28 +30,28 @@ class Apiinit
                 $verified = false;
             }
 
-            $last_verification = (int)get_option($module_name . '_last_verification');
-            $seconds = $data->check_interval ?? 0;
+            $last_verification = (int) get_option($module_name.'_last_verification');
+            $seconds           = $data->check_interval ?? 0;
 
             if (!empty($seconds) && time() > ($last_verification + $seconds)) {
                 $verified = false;
                 try {
                     $request = Flatadmintheme_Requests::post(VAL_PROD_POINT, ['Accept' => 'application/json', 'Authorization' => $token], json_encode(['verification_id' => $verification_id, 'item_id' => basename($module['headers']['uri']), 'activated_domain' => base_url()]));
-                    $status = $request->status_code;
+                    $status  = $request->status_code;
                     if ((500 <= $status && $status <= 599) || 404 == $status) {
-                        update_option($module_name . '_heartbeat', base64_encode(json_encode(['status' => $status, 'id' => $token, 'end_point' => VAL_PROD_POINT])));
+                        update_option($module_name.'_heartbeat', base64_encode(json_encode(['status' => $status, 'id' => $token, 'end_point' => VAL_PROD_POINT])));
                         $verified = false;
                     } else {
-                        $result = json_decode($request->body);
+                        $result   = json_decode($request->body);
                         $verified = !empty($result->valid);
                         if ($verified) {
-                            delete_option($module_name . '_heartbeat');
+                            delete_option($module_name.'_heartbeat');
                         }
                     }
                 } catch (Exception $e) {
                     $verified = false;
                 }
-                update_option($module_name . '_last_verification', time());
+                update_option($module_name.'_last_verification', time());
             }
         }
 
@@ -66,7 +65,7 @@ class Apiinit
 
     public static function ease_of_mind($module_name)
     {
-        if (!\function_exists($module_name . '_actLib') || !\function_exists($module_name . '_sidecheck') || !\function_exists($module_name . '_deregister')) {
+        if (!\function_exists($module_name.'_actLib') || !\function_exists($module_name.'_sidecheck') || !\function_exists($module_name.'_deregister')) {
             get_instance()->app_modules->deactivate($module_name);
         }
     }
@@ -74,13 +73,13 @@ class Apiinit
 
     public static function activate($module)
     {
-        if (!option_exists($module['system_name'] . '_verification_id') && empty(get_option($module['system_name'] . '_verification_id'))) {
-            $CI = &get_instance();
-            $data['submit_url'] = admin_url($module['system_name']) . '/env_ver/activate';
-            $data['original_url'] = admin_url('modules/activate/' . $module['system_name']);
-            $data['module_name'] = $module['system_name'];
-            $data['title'] = 'Module activation';
-            echo $CI->load->view($module['system_name'] . '/activate', $data, true);
+        if (!option_exists($module['system_name'].'_verification_id') && empty(get_option($module['system_name'].'_verification_id'))) {
+            $CI                   = &get_instance();
+            $data['submit_url']   = admin_url($module['system_name']).'/env_ver/activate';
+            $data['original_url'] = admin_url('modules/activate/'.$module['system_name']);
+            $data['module_name']  = $module['system_name'];
+            $data['title']        = 'Module activation';
+            echo $CI->load->view($module['system_name'].'/activate', $data, true);
             exit;
         }
     }
@@ -118,10 +117,10 @@ class Apiinit
         }
         $all_activated = get_instance()->app_modules->get_activated();
         foreach ($all_activated as $active_module => $value) {
-            $verification_id = get_option($active_module . '_verification_id');
+            $verification_id =  get_option($active_module.'_verification_id');
             if (!empty($verification_id)) {
                 $verification_id = base64_decode($verification_id);
-                $id_data = explode('|', $verification_id);
+                $id_data         = explode('|', $verification_id);
                 if ($id_data[3] == $code) {
                     return ['status' => false, 'message' => 'This Purchase code is Already being used in other module'];
                 }
@@ -135,39 +134,53 @@ class Apiinit
             'sold_at' => '2021-01-01 00:00:00',
         ];
 
-//        if (empty($envato_res) || !\is_object($envato_res) || isset($envato_res->error) || !isset($envato_res->sold_at)) {
-//            return ['status' => false, 'message' => 'Something went wrong'];
-//        }
-//        if (basename($module['headers']['uri']) != $envato_res->item->id) {
-//            return ['status' => false, 'message' => 'Purchase key is not valid'];
-//        }
-
+        if (empty($envato_res) || !\is_object($envato_res) || isset($envato_res->error) || !isset($envato_res->sold_at)) {
+            return ['status' => false, 'message' => 'Something went wrong'];
+        }
+        if (basename($module['headers']['uri']) != $envato_res->item->id) {
+            return ['status' => false, 'message' => 'Purchase key is not valid'];
+        }
         get_instance()->load->library('user_agent');
-        $data['user_agent'] = get_instance()->agent->browser() . ' ' . get_instance()->agent->version();
+        $data['user_agent']       = get_instance()->agent->browser().' '.get_instance()->agent->version();
         $data['activated_domain'] = base_url();
-        $data['requested_at'] = date('Y-m-d H:i:s');
-        $data['ip'] = self::getUserIP();
-        $data['os'] = get_instance()->agent->platform();
-        $data['purchase_code'] = $code;
-        $data['envato_res'] = $envato_res;
-        $data = json_encode($data);
+        $data['requested_at']     = date('Y-m-d H:i:s');
+        $data['ip']               = self::getUserIP();
+        $data['os']               = get_instance()->agent->platform();
+        $data['purchase_code']    = $code;
+        $data['envato_res']       = $envato_res;
+        $data                     = json_encode($data);
 
         try {
 
-            update_option($module_name . '_verification_id', base64_encode('TEST:'.$code));
-            update_option($module_name . '_last_verification', time());
-            update_option($module_name . '_product_token', base64_encode('TEST:'.$code));
-            delete_option($module_name . '_heartbeat');
+            //fake response
+            $response = [
+                'status_code' => 200,
+                'message' => "Activation successfully",
+                'data' => [
+                    'verification_id' => 'CX324-SADF23-ASDF23-ASDF32-234VDS',
+                    'token' => '_token_4234zxcva324adsf3234sd'
+                ]
+            ];
 
-            return ['status' => true];
 
+            $return = $response['data'] ?? [];
+
+            if (!empty($return)) {
+                update_option($module_name.'_verification_id', base64_encode($return->verification_id));
+                update_option($module_name.'_last_verification', time());
+                update_option($module_name.'_product_token', $return['token']);
+                delete_option($module_name.'_heartbeat');
+
+                return ['status' => true];
+            }
         } catch (Exception $e) {
-            update_option($module_name . '_verification_id', '');
-            update_option($module_name . '_last_verification', time());
-            update_option($module_name . '_heartbeat', base64_encode(json_encode(['status' => 200, 'id' => $code, 'end_point' => REG_PROD_POINT])));
+            update_option($module_name.'_verification_id', '');
+            update_option($module_name.'_last_verification', time());
+            update_option($module_name.'_heartbeat', base64_encode(json_encode(['status' => $request->status_code, 'id' => $code, 'end_point' => REG_PROD_POINT])));
 
             return ['status' => true];
         }
 
+        return ['status' => false, 'message' => 'Something went wrong'];
     }
 }
